requirements.txt
# 项目依赖

# OpenAI 兼容接口（支持ModelScope）
openai

# 命令行工具
click

# 配置管理
python-dotenv

# 代码处理
pygments

# 测试
pytest
